\name{multigraph-package}
\alias{multigraph-package}
%\alias{multigraph}
\docType{package}
\title{
Create and manipulate multigraphs
}
\description{
Functions to create and manipulate multigraphs
}
\details{
\tabular{ll}{
Package: \tab multigraph\cr
Type: \tab Package\cr
Version: \tab 0.30 (devel)\cr
Date: \tab 8 August 2016\cr
License: \tab GPL-3\cr
}
Here are the functions to create and manipulate multigraphs and bipartite multigraphs. Imports routines from the \code{multiplex} package version 1.9.9
}
\author{
J. Antonio Rivero Ostoic

Maintainer: Antonio Rivero Ostoic <jaro@econ.au.dk>
}
\keyword{ graphs}
\keyword{ manip }
\keyword{ IO }
\keyword{ data }
\keyword{ file }
