\name{multigraph-package}
\alias{multigraph-package}
%\alias{multigraph}
\docType{package}
\title{
Plot and Manipulate Multigraphs
}
\description{
Functions to create and manipulate multigraphs, bipartite graphs, and weighted multigraphs
}
\details{
\tabular{ll}{
Package: \tab multigraph \cr
Type: \tab Package \cr
Version: \tab 0.92 (devel) \cr
Depends: \tab multiplex (>= 2.9.2) \cr
Date: \tab 18 October 2019 \cr
License: \tab GPL-3 \cr
}
This package contains functions to create, plot and manipulate multigraphs, signed and weighted multigraphs, bipartite graphs, and Cayley graphs with multiple edges. 
%It depends on routines from the \code{multiplex} package with version 2.9 or higher.
}
\author{
J. Antonio Rivero Ostoic

Maintainer: Antonio Rivero Ostoic <multiplex@post.com>
}
\seealso{
\code{\link[multiplex:multiplex-package]{multiplex-package}}, \code{\link[multiplex:incubs]{incubs}}, \code{\link[multiplex:zbind]{zbind}}, \code{\link[multiplex:transf]{transf}}
}

\keyword{ graphs}
\keyword{ manip }
\keyword{ IO }
\keyword{ data }
\keyword{ file }
