\name{multigraph-package}
\alias{multigraph-package}
%\alias{multigraph}
\docType{package}
\title{
Plot and manipulate multigraphs
}
\description{
Functions to create and manipulate multigraphs
}
\details{
\tabular{ll}{
Package: \tab multigraph \cr
Type: \tab Package \cr
Version: \tab 0.30 (devel) \cr
Depends: \tab multiplex (>= 2.0) \cr
Date: \tab 16 August 2016 \cr
License: \tab GPL-3 \cr
}
This package contains functions to create, plot and manipulate multigraphs and bipartite graphs with multiple edges. It depends on routines from the \code{multiplex} package version 2.0
}
\author{
J. Antonio Rivero Ostoic

Maintainer: Antonio Rivero Ostoic <jaro@econ.au.dk>
}
\keyword{ graphs}
\keyword{ manip }
\keyword{ IO }
\keyword{ data }
\keyword{ file }
